feature: introduce `ModifiableClockTrait` and `ClockTrait`
By default, the abstract adapter tests do use the `ClockTrait` which uses `sleep` as before. This is necessary due to the fact that some adapters use dedicated services to persist cache values where we can't modify the time. Some cache adapters, such as `Memory` and `Filesystem`, persist the expiry timestamp and thus, the clock can be modified so that cache items are detected as expired.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>